+++ flexo-foundation/src/test/java/org/openflexo/foundation/OpenflexoTestCase.java
@@ -83,6 +83,12 @@ import org.openflexo.toolbox.FileUtils;
  */
 public abstract class OpenflexoTestCase {
 
+	/**
+	 * !!!!! IMPORTANT !!!!!<br>
+	 * Do not forget to set back this flag to true when committing into a production environment
+	 */
+	public static final boolean DELETE_TEST_RESOURCE_CENTER_AFTER_TEST_EXECUTION = true;
+
 	private static final Logger logger = FlexoLogger.getLogger(OpenflexoTestCase.class.getPackage().getName());
 
 	private static final String TEST_RESOURCE_CENTER_URI = "http://openflexo.org/test/TestResourceCenter";
@@ -117,12 +123,14 @@ public abstract class OpenflexoTestCase {
 	}
 
 	protected static void deleteTestResourceCenters() {
-		if (testResourceCenterDirectory != null) {
-			FileUtils.deleteDir(testResourceCenterDirectory);
-		}
-		if (testResourceCenterDirectoriesToRemove != null) {
-			for (File testResourceCenterDirectoryToRemove : testResourceCenterDirectoriesToRemove) {
-				FileUtils.deleteDir(testResourceCenterDirectoryToRemove);
+		if (DELETE_TEST_RESOURCE_CENTER_AFTER_TEST_EXECUTION) {
+			if (testResourceCenterDirectory != null) {
+				FileUtils.deleteDir(testResourceCenterDirectory);
+			}
+			if (testResourceCenterDirectoriesToRemove != null) {
+				for (File testResourceCenterDirectoryToRemove : testResourceCenterDirectoriesToRemove) {
+					FileUtils.deleteDir(testResourceCenterDirectoryToRemove);
+				}
 			}
 		}
 		resourceCenter = null;
